Fixes reorder with qml
